Rename metadata package -> metadatacache
