+++ src/zircon/metadatacache/metadata.go
@@ -1,4 +1,4 @@
-package metadata
+package metadatacache
 
 import (
 	"bytes"
