+++ src/zircon/metadatacache/metadata_test.go
@@ -1,4 +1,4 @@
-package metadata
+package metadatacache
 
 import (
 	"github.com/stretchr/testify/assert"
